split-utils * Move remaining utils functions to utils/utils.ts
