+++ src/containers/Admin/AdminPage.tsx
@@ -6,7 +6,7 @@ import { LockedTavle } from '../Error/ErrorPages'
 import { useUser } from '../../UserProvider'
 import { useSettings } from '../../settings/SettingsProvider'
 import { ThemeContrastWrapper } from '../ThemeWrapper/ThemeContrastWrapper'
-import { isDarkOrDefaultTheme } from '../../utils'
+import { isDarkOrDefaultTheme } from '../../utils/utils'
 import { LogoTab } from './LogoTab/LogoTab'
 import { EditTab } from './EditTab/EditTab'
 import { ThemeTab } from './ThemeTab/ThemeTab'
