+++ src/containers/Admin/EditTab/BikePanel/BikePanel.tsx
@@ -2,7 +2,7 @@ import React, { ChangeEvent, useCallback } from 'react'
 import { Station } from '@entur/sdk/lib/mobility/types'
 import { Checkbox, Fieldset } from '@entur/form'
 import { Paragraph } from '@entur/typography'
-import { getTranslation } from '../../../../utils'
+import { getTranslation } from '../../../../utils/utils'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { toggleValueInList } from '../../../../utils/array'
 import './BikePanel.scss'
