+++ src/containers/Admin/EditTab/EditTab.tsx
@@ -20,7 +20,7 @@ import { Station } from '@entur/sdk/lib/mobility/types'
 import { Button } from '@entur/button'
 import { Mode } from '../../../settings'
 import { useSettings } from '../../../settings/SettingsProvider'
-import { isMobileWeb, getTranslation } from '../../../utils'
+import { isMobileWeb, getTranslation } from '../../../utils/utils'
 import { DEFAULT_DISTANCE, DEFAULT_ZOOM } from '../../../constants'
 import { Line, StopPlaceWithLines } from '../../../types'
 import { useNearestPlaces, useRentalStations } from '../../../logic'
