+++ src/containers/Admin/EditTab/RealtimeDataPanel/RealTimeDataPanel.tsx
@@ -5,7 +5,7 @@ import { Loader } from '@entur/loader'
 import { Heading3, Label, Paragraph } from '@entur/typography'
 import { ExpandablePanel } from '@entur/expand'
 import { Line } from '../../../../types'
-import { transportModeNameMapper } from '../../../../utils'
+import { transportModeNameMapper } from '../../../../utils/utils'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { isTransport } from '../../../../utils/typeguards'
 import { getIcon } from '../../../../utils/icon'
