+++ src/dashboards/BusStop/MapTile/MapTile.tsx
@@ -3,7 +3,7 @@ import 'mapbox-gl/dist/mapbox-gl.css'
 import { Station, Vehicle } from '@entur/sdk/lib/mobility/types'
 import { Map } from '../../../components/Map/Map'
 import { StopPlaceWithDepartures } from '../../../types'
-import { isMobileWeb } from '../../../utils'
+import { isMobileWeb } from '../../../utils/utils'
 import './MapTile.scss'
 
 function MapTile(data: Props): JSX.Element {
