+++ src/dashboards/Chrono/BikeTile/BikeTile.tsx
@@ -5,7 +5,7 @@ import { Station } from '@entur/sdk/lib/mobility/types'
 import { Tile } from '../components/Tile/Tile'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { IconColorType } from '../../../types'
-import { getTranslation } from '../../../utils'
+import { getTranslation } from '../../../utils/utils'
 import { getIconColorType } from '../../../utils/icon'
 import { useWalkInfo, WalkInfo } from '../../../logic/use-walk-info/useWalkInfo'
 import { TileRow } from '../components/TileRow/TileRow'
