+++ src/dashboards/Compact/CompactDashboard.tsx
@@ -23,7 +23,7 @@ import {
 } from '../../settings/LocalStorage'
 import { useSettings } from '../../settings/SettingsProvider'
 import { DEFAULT_ZOOM, BREAKPOINTS } from '../../constants'
-import { isMobileWeb } from '../../utils'
+import { isMobileWeb } from '../../utils/utils'
 import { LongPressProvider } from '../../logic/longPressContext'
 import { WeatherTile } from '../../components/WeatherTile/WeatherTile'
 import { QRTile } from '../../components/QRTile/QRTile'
