+++ src/dashboards/Compact/components/TileRow/TileRow.tsx
@@ -6,7 +6,7 @@ import { TileSubLabel } from '../../../../types'
 import { ValidationExclamation } from '../../../../assets/icons/ValidationExclamation'
 import { ValidationError } from '../../../../assets/icons/ValidationError'
 import { SituationModal } from '../../../../components/SituationModal/SituationModal'
-import { isMobileWeb } from '../../../../utils'
+import { isMobileWeb } from '../../../../utils/utils'
 import { WalkInfo } from '../../../../logic/use-walk-info/useWalkInfo'
 import { PlatformInfo } from './PlatformInfo/PlatformInfo'
 import './TileRow.scss'
